MidiCCtoCV: processor post CC events as node messages.
UI uses that to display current value.

Project caches the value, and sends it back to the processor as the 'initial_value' in the spec, so
if the processor gets recreated (e.g. when ports change), it gets initialized with the latest known
value.

This does not yet solve the problem how to persist the current controller value across sessions.